+++ lib/src/json/editor_room.g.dart
@@ -13,6 +13,12 @@ EditorRoom _$EditorRoomFromJson(Map<String, dynamic> json) => EditorRoom(
   objects: (json['objects'] as List<dynamic>)
       .map((e) => EditorRoomObject.fromJson(e as Map<String, dynamic>))
       .toList(),
+  events: (json['events'] as List<dynamic>)
+      .map((e) => $enumDecode(_$AngstromEventTypeEnumMap, e))
+      .toList(),
+  eventComments: (json['eventComments'] as Map<String, dynamic>).map(
+    (k, e) => MapEntry($enumDecode(_$AngstromEventTypeEnumMap, k), e as String),
+  ),
   name: json['name'] as String? ?? 'Untitled Room',
   music: json['music'] == null
       ? null
@@ -22,13 +28,27 @@ EditorRoom _$EditorRoomFromJson(Map<String, dynamic> json) => EditorRoom(
   lastPageIndex: (json['lastPageIndex'] as num?)?.toInt() ?? 0,
 );
 
-Map<String, dynamic> _$EditorRoomToJson(EditorRoom instance) =>
-    <String, dynamic>{
-      'surfaces': instance.surfaces,
-      'objects': instance.objects,
-      'name': instance.name,
-      'music': instance.music,
-      'x': instance.x,
-      'y': instance.y,
-      'lastPageIndex': instance.lastPageIndex,
-    };
+Map<String, dynamic> _$EditorRoomToJson(
+  EditorRoom instance,
+) => <String, dynamic>{
+  'surfaces': instance.surfaces,
+  'objects': instance.objects,
+  'events': instance.events.map((e) => _$AngstromEventTypeEnumMap[e]!).toList(),
+  'eventComments': instance.eventComments.map(
+    (k, e) => MapEntry(_$AngstromEventTypeEnumMap[k]!, e),
+  ),
+  'name': instance.name,
+  'music': instance.music,
+  'x': instance.x,
+  'y': instance.y,
+  'lastPageIndex': instance.lastPageIndex,
+};
+
+const _$AngstromEventTypeEnumMap = {
+  AngstromEventType.onEnter: 'onEnter',
+  AngstromEventType.onMove: 'onMove',
+  AngstromEventType.onExit: 'onExit',
+  AngstromEventType.onApproach: 'onApproach',
+  AngstromEventType.onActivate: 'onActivate',
+  AngstromEventType.onLeave: 'onLeave',
+};